front boardlikebutton

+++ tripfront/src/api/boardapi.js
@@ -38,14 +38,14 @@ async function commentUpdate(content_id, comment, success, fail) {
 async function commentDelete(content_id, comment, success, fail) {
   await api.delete(`/board/${content_id}/comment`, JSON.stringify(comment)).then(success).catch(fail);
 }
-async function likeInsert(content_id, success, fail) {
-  await api.get(`/board/${content_id}/like`).then(success).catch(fail);
+async function likeSelect(content_id, user_id, success, fail) {
+  await api.get(`/board/${content_id}/like?user_id=${user_id}`).then(success).catch(fail);
 }
-async function likeSelect(content_id, success, fail) {
-  await api.post(`/board/${content_id}/like`).then(success).catch(fail);
+async function likeInsert(content_id, user_id, success, fail) {
+  await api.post(`/board/${content_id}/like?user_id=${user_id}`).then(success).catch(fail);
 }
-async function likeDelete(content_id, success, fail) {
-  await api.delete(`/board/${content_id}/like`).then(success).catch(fail);
+async function likeDelete(content_id, user_id, success, fail) {
+  await api.delete(`/board/${content_id}/like?user_id=${user_id}`).then(success).catch(fail);
 }
 
 export {
